add first integration tests
