+++ integration-tests/cra5/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@apollo/client": "3.13.0",
+    "@apollo/client": "4.1.3",
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "graphql": "^16.8.1",
@@ -11,7 +11,7 @@
     "react-dom": "^18.3.0",
     "react-scripts": "5.0.1",
     "rxjs": "^7.8.2",
-    "typescript": "^4.9.5"
+    "typescript": "^5.9.3"
   },
   "scripts": {
     "start": "PORT=3000 react-scripts start",
