+++ integration-tests/next/package.json
@@ -11,7 +11,7 @@
     "test": "playwright test"
   },
   "dependencies": {
-    "@apollo/client": "3.13.0",
+    "@apollo/client": "4.1.3",
     "@apollo/client-integration-nextjs": "^0.12.0-alpha.1",
     "@graphql-tools/schema": "^10.0.0",
     "@types/node": "20.3.1",
@@ -25,7 +25,7 @@
     "react-dom": "^18.3.0",
     "rxjs": "^7.8.2",
     "ts-invariant": "^0.10.3",
-    "typescript": "5.1.3"
+    "typescript": "5.9.3"
   },
   "devDependencies": {
     "@playwright/test": "*",
