+++ integration-tests/node/package.json
@@ -10,7 +10,7 @@
     "test:esm": "node --test test-esm.mjs"
   },
   "dependencies": {
-    "@apollo/client": "3.13.0",
+    "@apollo/client": "4.1.3",
     "react": "^18.3.0",
     "react-dom": "^18.3.0"
   },
