+++ integration-tests/type-tests/.gitignore
@@ -1,0 +1,2 @@
+*.js
+*.d.ts
