+++ integration-tests/type-tests/defaultOptions/defaults/index.ts
@@ -1,0 +1,7 @@
+import { InMemoryCache } from "@apollo/client";
+import { ApolloClient, ApolloLink } from "@apollo/client";
+
+new ApolloClient({
+  link: ApolloLink.empty(),
+  cache: new InMemoryCache(),
+});
