+++ integration-tests/type-tests/defaultOptions/query/errorPolicyAll/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+  }
+}
