+++ integration-tests/type-tests/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true,
+  },
+  "references": [
+    {
+      "path": "./defaultOptions/defaults/tsconfig.json"
+    },
+    {
+      "path": "./defaultOptions/query/errorPolicyAll/tsconfig.json"
+    }
+  ]
+}
