+++ integration-tests/vite-swc/package.json
@@ -12,7 +12,7 @@
     "test": "playwright test"
   },
   "dependencies": {
-    "@apollo/client": "3.13.0",
+    "@apollo/client": "4.1.3",
     "react": "^18.3.0",
     "react-dom": "^18.3.0",
     "rxjs": "^7.8.2"
@@ -29,7 +29,7 @@
     "eslint-plugin-react-refresh": "^0.3.4",
     "playwright": "*",
     "shared": "workspace:^",
-    "typescript": "^5.0.2",
+    "typescript": "^5.9.3",
     "vite": "^6.2.0"
   },
   "overrides": {
