+++ integration-tests/vite/package.json
@@ -12,7 +12,7 @@
     "test": "playwright test"
   },
   "dependencies": {
-    "@apollo/client": "3.13.0",
+    "@apollo/client": "4.1.3",
     "expect-type": "^1.2.1",
     "graphql": "^16.8.1",
     "react": "^18.3.0",
@@ -31,7 +31,7 @@
     "eslint-plugin-react-refresh": "^0.3.4",
     "playwright": "*",
     "shared": "workspace:^",
-    "typescript": "^5.0.2",
+    "typescript": "^5.9.3",
     "vite": "^4.3.9"
   }
 }
