--- integration-tests/vite/src/defaultOptions.ts
@@ -1,84 +1,0 @@
-import { InMemoryCache } from "@apollo/client";
-import { ApolloClient, ApolloLink } from "@apollo/client";
-
-declare module "@apollo/client" {
-  namespace ApolloClient {
-    namespace DeclareDefaultOptions {
-      interface WatchQuery {}
-      interface Query {
-        errorPolicy: "none";
-      }
-      interface Mutate {}
-    }
-  }
-}
-if (2 < 1 /* skip running this, it's just a type test */) {
-  // @ts-expect-error: Property 'defaultOptions' is missing in type '{ link: ApolloLink; cache: InMemoryCache; }' but required in type 'Options'.
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-  });
-
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-    // @ts-expect-error: Property 'query' is missing in type '{}' but required in type 'DefaultOptions'.
-    defaultOptions: {},
-  });
-
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-    defaultOptions: {
-      // @ts-expect-error: Property 'errorPolicy' is missing in type '{}' but required in type ...
-      query: {},
-    },
-  });
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-    defaultOptions: {
-      query: {
-        // @ts-expect-error: Type '"all"' is not assignable to type '"none"'.
-        errorPolicy: "all",
-      },
-    },
-  });
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-    defaultOptions: {
-      watchQuery: {
-        // @ts-expect-error: Type '"none"' is not assignable to type '"A default option for watchQuery.errorPolicy must be declared in ApolloClient.DeclareDefaultOptions before usage. See <TODO documentation link>."'.
-        errorPolicy: "none",
-      },
-      query: {
-        errorPolicy: "none",
-      },
-      mutate: {
-        // @ts-expect-error: Type '"none"' is not assignable to type '"A default option for mutate.errorPolicy must be declared in ApolloClient.DeclareDefaultOptions before usage. See <TODO documentation link>."'.
-        errorPolicy: "none",
-      },
-    },
-  });
-
-  new ApolloClient({
-    link: ApolloLink.empty(),
-    cache: new InMemoryCache(),
-    defaultOptions: {
-      watchQuery: {
-        context: {
-          headers: {
-            "x-custom-header": "custom-value",
-          },
-        },
-      },
-      query: {
-        errorPolicy: "none",
-      },
-      mutate: {
-        awaitRefetchQueries: true,
-      },
-    },
-  });
-}
