+++ integration-tests/wrapping-library/package.json
@@ -1,10 +1,10 @@
 {
   "name": "wrapping-library",
   "devDependencies": {
-    "typescript": "^5.7.3"
+    "typescript": "^5.9.3"
   },
   "dependencies": {
-    "@apollo/client": "3.13.0"
+    "@apollo/client": "4.1.3"
   },
   "scripts": {
     "build": "tsc",
